ControlClasses: Add optional callback argument to all methods
This commit introduces an optional callback based variant of all methods
in control classes. This variant accepts a result callback which
receives two arguments:

- a boolean flag which is true on success and false on error
- the result which is either the return value(s) or an error

The purpose of this callback based API is to work around situations in
which subscriptions to events need to be reliably combined with method
calls. Due to the fact that promises introduce additional delays, events
and method calls can race against each other.