UDPConnection: Improve estimation of sentTime
Each pending command keeps track of the time at which it has been sent.
This information is used to determine at which time it will be necessary
to retry the same command in unreliable transports. The sentTime did not
take the current packet queue length into consideration, which could
lead to it being very inaccurate in situations in which the send queue
was large.

This commit includes the send queue length into the calculation, which
can prevent commands from spuriously timing out.