Events: Emit event data as specified
The AES70 standard specifies each event with a corresponding event
type. Previously, events were emitted with the list of properties
of these event datatypes. This commit changes that and aligns the
APIs with the standard. Events are emitted with the datatype which
is used to define them in the standard.