RemoteDevice: Fix role name collision handling
When creating a rolemap for a device in which two objects have
the same role name and also an objects exists which has that role name
with a numeral suffix the tree_to_rolemap function would end in an
infinite loop.

This commit also adds some tests to check that the handling of that
situation works correctly.