RemoteDevice: Add stack debug feature
When enabled (using RemoteDevice.enable_stack_debug) stack traces are
collected when a command is sent. Then, if the device responds with
an error, the created RemoteError has the collected stack attached.

Note that this feature only works in javascript engines which implement
the Error stack property as propsed in
https://github.com/tc39/proposal-error-stacks