Refactor UDP command retry handling
The new strategy is to write commands
to the send buffer only until a small number
of write operations are pending.

Commands are assigned their txTime when
this happens. The benefit of this is that a
large write buffer will not lead to spurious
command timouts.

This new strategy also makes sure that command
messages are batched more efficiently.